make vcf2tsv script more general
