+++ data/Makefile
@@ -181,11 +181,17 @@ common_input/oncokb_cancer_genes_list_from_API.json:
 	curl "https://www.oncokb.org/api/v1/utils/cancerGeneList" -H "accept: application/json" -H "Authorization: Bearer $(ONCOKB_TOKEN)" | python -m json.tool > $@
 
 # download GRCh37 ClinVar VCF file from NCBI
-clinvar/input/clinvar_grch37.vcf:
+clinvar/input/clinvar_grch37_input.vcf:
 	curl "ftp://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh37/clinvar.vcf.gz" | gunzip -c > $@
-# download GRCh39 ClinVar VCF file from NCBI
-clinvar/input/clinvar_grch38.vcf:
+# generate GRCh37 ClinVar tsv file
+clinvar/export/clinvar_grch37.txt.gz:
+	python ../scripts/transform_vcf_to_tsv.py ../data/clinvar/input/clinvar_grch37_input.vcf ../data/clinvar/export/clinvar_grch37.txt --fields chromosome,start_position,end_position,reference_allele,alternate_allele,clinvar_id,clnsig,clnsigconf
+# download GRCh38 ClinVar VCF file from NCBI
+clinvar/input/clinvar_grch38_input.vcf:
 	curl "ftp://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh38/clinvar.vcf.gz" | gunzip -c > $@
+# generate GRCh38 ClinVar tsv file
+clinvar/export/clinvar_grch38.txt.gz:
+	python ../scripts/transform_vcf_to_tsv.py ../data/clinvar/input/clinvar_grch38_input.vcf ../data/clinvar/export/clinvar_grch38.txt --fields chromosome,start_position,end_position,reference_allele,alternate_allele,clinvar_id,clnsig,clnsigconf
 
 # create directories if not extistent
 dirs: $(VERSION)/input $(VERSION)/export
